+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+Zarcel v0.3.0 *(2018-08-23)*
+----------------------------
+* Thêm SerializedByteBufferInput và SerializedByteBufferOutput vào zarcel-core.
+
 Zarcel v0.2.0 *(2018-08-13)*
 ----------------------------
 * Thêm @Ignore để bỏ qua warning cho thuộc tính private.